Finally - Pet emotions are fully added!
the emotions are now more cohesive in the code and simpler! I should probably go fix up the other ones, but a small part of me is afraid Ill break it

I've got to clean up the item database and make objects and items more seperated as a whole. But I believe thats the end to it? I hope so.

I have to debug death sprites and such but lets see